rewrote the wasm benchmark. updated readme
