+++ neon/native/Cargo.toml
@@ -15,4 +15,4 @@ neon-build = "0.3"
 
 [dependencies]
 neon = "0.3"
-sha1 = "0.2.0"
+sha1 = "0.6.0"
